tests/wpt/web-platform-tests/battery-status/battery-promise.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Battery Test: navigator.getBattery() always return same battery promise</title>
<link rel="author" title="YuichiNukiyama" href="https://github.com/YuichiNukiyama">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
promise_test(function () {
  return navigator.getBattery().then(function (result) {
    assert_class_string(result, 'BatteryManager',
        'getBattery should return BatteryManager Object.');
  });
}, 'navigator.getBattery() shall return BatteryManager as a promise');

test(function () {
  assert_equals(navigator.getBattery(), navigator.getBattery());
}, 'navigator.getBattery() shall always return the same promise');
</script>
